added transaction

+++ src/database/080_security_question.sql
@@ -1,5 +1,5 @@
 CREATE TABLE security_question (
-     security_question_id  MEDIUMINT NOT NULL AUTO_INCREMENT,
+     security_question_id  INT NOT NULL AUTO_INCREMENT,
      question              VARCHAR(2000) NOT NULL,
      retired               DATETIME NOT NULL,
      PRIMARY KEY (security_question_id)
