+++ src/database/100_service.sql
@@ -1,5 +1,5 @@
 CREATE TABLE service (
-     service_id         MEDIUMINT NOT NULL AUTO_INCREMENT,
+     service_id         INT NOT NULL AUTO_INCREMENT,
      service_name       VARCHAR(255) NOT NULL,
      cost               DECIMAL(10,2) NOT NULL,
      retired            DATETIME NOT NULL,
